Extract classes originally in SettingsPage
